--- a/benchmark/vm_iclass_super.yml
+++ b/benchmark/vm_iclass_super.yml
@@ -1,0 +1,20 @@
+prelude: |
+  class C
+    def m
+      1
+    end
+
+    ("A".."M").each do |module_name|
+      eval <<-EOM
+          module #{module_name}
+            def m; super; end
+          end
+          prepend #{module_name}
+      EOM
+    end
+  end
+
+  obj = C.new
+benchmark:
+  vm_iclass_super: obj.m
+loop_count: 6000000