@ko1 2020-09-18
I'm not sure it can work on corner cases, for example using `define_method` to share the same iseq with different classes.