@eileencodes 2020-09-22
@tenderlove I fixed the benchmark. I also updated the benchmarks in my commit message because it seems like change was introduced to Ruby recently that made it even faster. I'm going to search for which commit after I finish writing this comment.

@ko1 - Aaron and I looked at your concerns today and determined this change is safe.

The call cache is not a problem here because:

1) we have a type check so there's no issues with lambda's
2) `define_method` can only be called on classes in the same inheritance chain
3) we're not skipping invalidating the call cache, just skipping the is_a? check

```ruby
class A
  def bar; :bar; end 
end

class B < A 
  def bar; super; end 
end

class C
  define_method(:baz, A.instance_method(:bar))
end

p C.new.baz
=> ruby_scripts/ruby_file.rb:9:in `define_method': bind argument must be a subclass of A (TypeError)
	from ruby_scripts/ruby_file.rb:9:in `<class:C>'
	from ruby_scripts/ruby_file.rb:8:in `<main>'

```