@tenderlove 2020-09-22
The iseq pointer comparison is [restricted to only methods iseq type methods](https://github.com/ruby/ruby/pull/3545/files#diff-347392b62c9edab3c8ab5fb0d71a012fR680) (*not* block methods). `define_method` only allows iseq methods to be defined on objects with the the same ancestors.  I think this means that if the current frame's iseq is the same as the defined method iseq, we can skip the `is_a?` check because the `is_a?` check was made at _method definition time_.