@ko1 2020-09-23
How about that?

```
module M
  def foo
    super
  end
end

class C0
  def foo() = :C0
end
class D0
  def foo() = :D0
end

class C1 < C0
  define_method(:foo, M.instance_method(:foo))
end
class D1 < D0
  define_method(:foo, M.instance_method(:foo))
end

p C1.new.foo
p D1.new.foo
```