@eileencodes 2020-09-23
On master and on this branch I get:

```
:C0
:D0
```

On Ruby 2.7 I get:

```
ruby_scripts/ko1_example.rb:8: syntax error, unexpected '='
  def foo() = :C0
ruby_scripts/ko1_example.rb:11: syntax error, unexpected '='
  def foo() = :D0
ruby_scripts/ko1_example.rb:22: syntax error, unexpected end-of-input, expecting `end'
```