@tenderlove 2020-09-23
@ko1 on master:

```
[aaron@tc-lan-adapter ~/g/ruby (master)]$ make runruby
./miniruby -I./lib -I. -I.ext/common  ./tool/runruby.rb --extout=.ext  -- --disable-gems ./test.rb
:C0
:D0
```

On this branch:

```
[aaron@tc-lan-adapter ~/g/ruby (improve-performance-of-super)]$ make runruby
./miniruby -I./lib -I. -I.ext/common  ./tool/runruby.rb --extout=.ext  -- --disable-gems ./test.rb
:C0
:D0
```

This change has no impact to the inline cache.  We found that [this line](https://github.com/ruby/ruby/pull/3545/files#diff-347392b62c9edab3c8ab5fb0d71a012fL3302) is a bottleneck.  We can skip that check if we compare iseq pointers.  That conditional is to check [instance_eval inside a method](https://github.com/ruby/ruby/blob/b717f73402dc6f2d6ba6a1d72d9c8543a4f2cd27/test/ruby/test_super.rb#L276-L283), and we can skip the check by comparing the control frame to the method definition.

@eileencodes this syntax is new in master, you need to compare master to this branch 😁